Docker/telegram_bot/requirements.txt
python-telegram-bot==20.3
pandas
fpdf